refactor(codegen): split into 2 commands
Generates the strapi schema file in one command, then runs the command again in watch mode with the primary configuration.
